[2-Tuesday] Working through solution code for the react / spring boot connection lab.

+++ Week10/Lessons/2-Tuesday/4-React-Spring-Boot-Lab/spring-boot-monolith/src/main/java/com/example/springbootmonolith/repositories/CourseRepository.java
@@ -4,5 +4,5 @@ import com.example.springbootmonolith.models.Course;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CourseRepository extends CrudRepository<Course, Integer> {
-    Course findById(long id);
+    Course findById(int id);
 }
